refactor(favorites): split provider file into 3 — one class per file (#727)
favorites_provider.dart had grown to 423 LOC, the first entry on
#727's hit list. Split into three sibling files, no behaviour
change.

- `favorites_provider.dart` (213 LOC) keeps the `Favorites`
  keepAlive notifier (add/remove/toggle/_reload/_refreshWidget)
- `favorite_stations_provider.dart` (231 LOC) owns the
  `FavoriteStations` async notifier — the per-country fetch +
  price-refresh logic that was the bulk of the original file
- `is_favorite_provider.dart` (18 LOC) holds the two one-liner
  helper providers

Re-exports at the top of `favorites_provider.dart` keep every
existing `import '.../favorites_provider.dart'` working without a
single call-site change.

All 105 favorites + favorites-store tests pass; 0 analyze issues.

Part of #727 (first of 4 files split — search_provider.dart,
sync_service.dart, add_fill_up_screen.dart still pending).
